Moving unit tests to the correct package
