+++ mezzanine-compiler/com/anthonycr/mezzanine/MezzanineGenerator.java
@@ -1,0 +1,16 @@
+package com.anthonycr.mezzanine;
+
+import java.lang.Override;
+import java.lang.String;
+
+public final class MezzanineGenerator {
+  private MezzanineGenerator() {
+  }
+
+  public static class ValidFileTestCase implements testcase.ValidFileTestCase {
+    @Override
+    public String test() {
+      return "{\n  \"field\": \"value\"\n}";
+    }
+  }
+}
